+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
     Fixed -- for any bug fixes.
     Security -- in case of vulnerabilities.
 -->
+## [0.15.1]
+
+
 
 ## [0.15.0]
 
@@ -64,7 +67,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Debugger is implemented
 
 
-[Unreleased]: https://github.com/TEK-Engineering/tsp-toolkit-kic-debug/compare/v0.15.0...HEAD
+[Unreleased]: https://github.com/TEK-Engineering/tsp-toolkit-kic-debug/compare/v0.15.1...HEAD
+[0.15.1]: https://github.com/TEK-Engineering/tsp-toolkit-kic-debug/releases/tag/v0.15.1
 [0.15.0]: https://github.com/TEK-Engineering/tsp-toolkit-kic-debug/releases/tag/v0.15.0
 [0.14.1]: https://github.com/TEK-Engineering/tsp-toolkit-kic-debug/releases/tag/v0.14.1
 [0.13.2]: https://github.com/TEK-Engineering/tsp-toolkit-kic-debug/releases/tag/v0.13.2